# swagger.yaml
openapi: 3.0.3
info:
  title: SLSEA Solar Generation Data API
  description: >
    Real-time and historical solar generation data API for the Sri Lanka
    Sustainable Energy Authority. Metering devices push readings as write
    clients; SLSEA users read data scoped by jurisdiction.
  version: 1.0.0
servers:
  - url: https://webapicw-production.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development
...
paths:
  /provinces:
    get:
      tags: [Hierarchy]
      summary: List all provinces
      responses:
        '200':
          description: Array of provinces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Province'
        '406':
          $ref: '#/components/responses/NotAcceptable'
...
  /provinces/{id}/districts:
    get:
      tags: [Hierarchy]
      summary: List districts in a province (scoped collection)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Array of districts in this province
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/District'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /districts:
    get:
      tags: [Hierarchy]
      summary: List districts (scoped to the logged-in user's jurisdiction)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of districts visible to this user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistrictDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /districts/{id}:
    get:
      tags: [Hierarchy]
      summary: Get a single district (atomic resource)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: One district (with populated province)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /installations:
    get:
      tags: [Installations]
      summary: List installations (scoped to the logged-in user's jurisdiction)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of installations visible to this user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Installation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
    post:
      tags: [Installations]
      summary: Provision a new installation (returns a one-time device API key)
      description: >
        Creates an installation and generates its device API key. The plain key
        is returned ONCE; only its hash is stored. The device uses that key at
        POST /installations/{id}/token to obtain write tokens.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [meterId, capacityKw, latitude, longitude, substation]
              properties:
                meterId: { type: string, minLength: 1, maxLength: 50 }
                capacityKw: { type: number, exclusiveMinimum: true, minimum: 0 }
                latitude: { type: number, minimum: -90, maximum: 90 }
                longitude: { type: number, minimum: -180, maximum: 180 }
                substation: { type: string, description: "Substation ObjectId" }
      responses:
        '201':
          description: Installation created (apiKey shown once)
          headers:
            Location:
              description: URL of the newly created installation
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Installation'
                  - type: object
                    properties:
                      apiKey:
                        type: string
                        description: Plain device API key — shown only once
        '400':
          description: Validation failed (VALIDATION_ERROR, field list in error.detail) or unknown substation (INVALID_SUBSTATION)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: An installation with this meterId already exists
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: National-role access required
...
  /installations/{id}/readings:
    get:
      tags: [Readings]
      summary: Paginated, filterable, sortable reading history (analytical resource)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: from
          in: query
          description: ISO-8601 date or date-time — start of time window filter
          schema: { type: string, format: date }
        - name: to
          in: query
          description: ISO-8601 date or date-time — end of time window filter (must not be earlier than `from`)
          schema: { type: string, format: date }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: Paginated readings with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reading'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Installation is outside your jurisdiction
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Readings]
      summary: Device pushes a new generation reading (write path, JWT-authenticated)
      description: >
        Requires a DEVICE token (installation-write scope) bound to this
        installation. SLSEA user tokens are rejected with 403 WRONG_TOKEN_TYPE.
        The installation must still exist (404 otherwise).
      security:
        - deviceBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [powerKw, energyKwh, voltage]
              properties:
                timestamp:
                  type: string
                  format: date-time
                  description: ISO-8601. Defaults to the server time. May not be more than 5 minutes in the future.
                powerKw: { type: number, minimum: 0 }
                energyKwh: { type: number, minimum: 0, description: Cumulative energy meter value }
                voltage: { type: number, minimum: 0 }
      responses:
        '201':
          description: Reading created
          headers:
            Location:
              description: URL of the newly created reading
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            WRONG_TOKEN_TYPE (a user token was used), WRONG_SCOPE, or
            WRONG_INSTALLATION (token is bound to a different installation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        SLSEA USER token from POST /auth/login (valid 8 hours). Used for every
        read/admin endpoint. A device token is rejected here with
        403 WRONG_TOKEN_TYPE, even though it is validly signed.
    deviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        DEVICE token from POST /installations/{id}/token (valid 1 hour, scope
        installation-write, bound to one installation). Used only to push
        readings. A user token is rejected here with 403 WRONG_TOKEN_TYPE.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Device API key, issued at installation provisioning
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
      description: MongoDB ObjectId (24 hex characters). A malformed value is a 400 INVALID_ID.
  responses:
    ValidationError:
      description: >
        Malformed id (INVALID_ID), invalid JSON (INVALID_JSON) or a field that failed
        validation (VALIDATION_ERROR). For VALIDATION_ERROR and INVALID_ID, error.detail
        is a list of `{ location, field, message }` entries — one per problem.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Request body exceeds the size limit (PAYLOAD_TOO_LARGE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: Request body was not sent as application/json (UNSUPPORTED_MEDIA_TYPE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotAcceptable:
      description: The requested representation is not available; this API only produces application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Province:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
    District:
      type: object
      description: Shape returned by GET /provinces/{id}/districts — province is just an ObjectId reference
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        province: { type: string, description: "ObjectId reference to the parent province" }
    DistrictDetail:
      type: object
      description: Shape returned by GET /districts — province is populated
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        province:
          type: object
          properties:
            _id: { type: string }
            name: { type: string }
            code: { type: string }
...
    Installation:
      type: object
      description: Shape returned by GET /installations (list view)
      properties:
        _id: { type: string }
        meterId: { type: string }
        capacityKw: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        substation:
          type: object
          properties:
            _id: { type: string }
            name: { type: string }
            code: { type: string }
...
    Reading:
      type: object
      properties:
        _id: { type: string }
        installation: { type: string }
        timestamp: { type: string, format: date-time }
        powerKw: { type: number }
        energyKwh: { type: number }
        voltage: { type: number }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
        nextPage: { type: string, nullable: true }
        prevPage: { type: string, nullable: true }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            detail: { nullable: true }